+++ src/components/legends/berthLegend/selectedBerthsLegend/selectedBerthsLegend.scss
@@ -1,0 +1,6 @@
+@import 'helsinki/colors';
+
+.vene-berths-legend__legend {
+  background-color: $hel-fog;
+  padding-bottom: 3em;
+}
