+++ src/components/legends/formLegend/formLegend.scss
@@ -1,0 +1,3 @@
+.vene-form-legend {
+  padding-bottom: 2em;
+}
