+++ src/components/pages/formPage/FormPage.tsx
@@ -3,11 +3,12 @@ import { Col, Container, Row } from 'reactstrap';
 
 import Wizard from '../../forms/Wizard';
 import Layout from '../../layout/Layout';
-import FormLegend from '../../legends/FormLegend';
+import FormLegend from '../../legends/formLegend/FormLegend';
 import Steps from '../../steps/StepsContainer';
 
 import { StepType } from '../../steps/step/Step';
-import './FormPage.scss';
+
+import './formPage.scss';
 
 interface Props {
   initialValues: {};
