+++ src/components/pages/formPage/formPage.scss
@@ -1,0 +1,5 @@
+@import 'helsinki/colors';
+
+.vene-form-page {
+  background-color: $hel-fog;
+}
